Improve eating behaviour, implementing it as a competence

+++ src/main/scala/model/Food.scala
@@ -2,31 +2,25 @@ package model
 
 import utility.Geometry.Vector2D
 
-/**
- * This class represents a food source.
- * */
-case class Food(override val position: Vector2D, quantity: Int)
-  extends SimpleObstacle(position, math.log(quantity)*10, math.log(quantity)*10) {
+/** A food source. */
+case class Food(override val position: Vector2D, quantity: Double)
+  extends SimpleObstacle(position, math.log(quantity) * 10, math.log(quantity) * 10) {
 
-  /**
-   * Increase food quantity.
+  /** Increase food quantity.
    *
-   * @param newQuantity to increase at actual quantity
-   *
+   * @param newQuantity to increase actual quantity
    * @return new instance of Food with increased quantity
    * */
-  def + (newQuantity: Int): Food = {
+  def + (newQuantity: Double): Food = {
     Food(position, quantity + newQuantity)
   }
 
-  /**
-   * Decrease food quantity.
+  /** Decrease food quantity.
    *
-   * @param newQuantity to increase at actual quantity
-   *
+   * @param newQuantity to decrease actual quantity
    * @return new instance of Food with decreased quantity
    * */
-  def - (newQuantity: Int): Food = {
+  def - (newQuantity: Double): Food = {
     this + ( - newQuantity)
   }
 }
